#12
As a work around resigning/defeated players becoming observers will receive the Espionage tech. They will then be able to see various stats again.

It's a little weird because being an observer at the start of the game is different from a player turning into an observer. Because in the last case the player keep his playerID. It's probably done this way so that the system knows who to give certain statistics to at the summary and ratings and such.