massive template rework
from now on only modify what i need in a template and send it back to vanilla parent. This minimizes the risk of needing to scan an entire template for errors in future alphas. If there is something new to be merged it'll just simply take it from vanilla parent.

+ template fixes part 2